pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "warehouse-management-system"
version = "0.1.0"
description = "FastAPI + SQLite Warehouse Management System API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
  "sqlalchemy>=2.0.38,<3.0.0",
  "pydantic-settings>=2.8.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
  "alembic>=1.14.1,<2.0.0",
  "pytest>=8.3.4,<9.0.0",
  "httpx>=0.28.1,<1.0.0",
  "ruff>=0.9.7,<1.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.hatch.build.targets.wheel]
packages = ["app"]
